requirements.txt
Django==4.0.7
djangorestframework==3.13.1
django-cors-headers==3.13.0
pytest==7.1.3
pytest-django==4.5.2
numexpr==2.10.2